Main

Documentation

window :: Display

Gloss window configuration.

loadJuicyPNGS

Arguments

:: [FilePath]
Sprite picture paths

-> [IO (Maybe Picture)]

Load multiple PNGs with gloss-juicy

listFromIO :: (Traversable t, Monad m) => t (m a) -> m (t a)

loadSprites :: IO Sprites

Load all sprites from files into Sprites.

saveLevel

Arguments

:: FilePath
Path where to save level

-> MapInfo
Level to save

-> IO ()

Save level by path.

loadAllSequentialLevelsFrom

Arguments

:: Int
Starting index

-> IO [MapInfo]
List of levels from starting sindex until levels cannot be loaded.

Load levels sequentially from some index. If a level cannot be loaded, the procedure is stopped and level list is returned.

loadLevel

Arguments

:: FilePath
Level path

-> IO (Maybe MapInfo)

Try load level by path.

loadLevelIfUnloaded

Arguments

:: MetaInfo
Level

-> IO (MetaInfo, MapInfo)

Try load the level only if preloaded levels have been exhausted. If level cannot be loaded, load empty map. Return new level and updated metaInfo.

data GameState

Game state that represents if the game is in MapEditor or in Play Mode.

Constructors

GameOn GameState

EditorOn MapEditorState

main :: IO ()

getLevelPath

Arguments

:: Int
Level index

-> FilePath
Level path

Convert level index to filePath

addSoundToRequests

Arguments

:: String
Sound name

-> MetaInfo
Input metaInfo

-> MetaInfo
Updated metaInfo

Update sound requests by adding one sounds, and return new MetaInfo